+++ src/main/java/com/tnecnivvv/lobbySystem/plugin/PluginHelper.java
@@ -1,7 +1,15 @@
 package com.tnecnivvv.lobbySystem.plugin;
 
 import com.tnecnivvv.api.config.ConfigurationAPI;
+import com.tnecnivvv.api.mysql.DatabaseAPI;
+import com.tnecnivvv.api.mysql.redis.Redis;
+import com.tnecnivvv.api.mysql.source.DataSource;
 import com.tnecnivvv.lobbySystem.Registry;
+import com.tnecnivvv.lobbySystem.dependencies.SimpleLogger;
+import com.tnecnivvv.lobbySystem.dependencies.TextSerializer;
+import com.tnecnivvv.lobbySystem.dependencies.Updater;
+
+import java.util.logging.Level;
 
 public class PluginHelper {
 
@@ -28,11 +36,40 @@ public class PluginHelper {
         return instance;
     }
 
-    public void initialize() {
+    public void initializeHelper() {
         registry.registerClass(ConfigurationAPI.class, new ConfigurationAPI());
+        registry.registerClass(SimpleLogger.class, new SimpleLogger());
+
+        registry.registerClass(DataSource.class, new DataSource());
+        registry.registerClass(Redis.class, new Redis());
+        registry.registerClass(DatabaseAPI.class, new DatabaseAPI());
+
+        registry.registerClass(Updater.class, new Updater());
+        registry.registerClass(TextSerializer.class, new TextSerializer());
+    }
+
+    public void initializeDependencies() {
+        try {
+            registry.resolveClass(ConfigurationAPI.class).saveConfiguration();
+            registry.resolveClass(DatabaseAPI.class).initialize();
+
+            Runnable updaterTasks = () -> {
+                Updater updater = registry.resolveClass(Updater.class);
+                updater.checkOnStart();
+                updater.checkForNotification();
+            };
+            updaterTasks.run();
+        } catch (Exception exception) {
+            registry.resolveClass(SimpleLogger.class).log(Level.SEVERE, "Error initializing dependencies", exception);
+        }
     }
 
     public void terminate() {
-        registry.clearCache();
+        try {
+            registry.resolveClass(ConfigurationAPI.class).clearAllCache();
+            registry.resolveClass(DatabaseAPI.class).terminate();
+        } finally {
+            registry.clearCache();
+        }
     }
 }