+++ src/main/java/com/tnecnivvv/lobbySystem/plugin/ServiceLocator.java
@@ -1,7 +1,11 @@
 package com.tnecnivvv.lobbySystem.plugin;
 
 import com.tnecnivvv.api.config.ConfigurationAPI;
+import com.tnecnivvv.api.mysql.DatabaseAPI;
+import com.tnecnivvv.api.mysql.redis.Redis;
+import com.tnecnivvv.api.mysql.source.DataSource;
 import com.tnecnivvv.lobbySystem.Registry;
+import com.tnecnivvv.lobbySystem.dependencies.SimpleLogger;
 
 public class ServiceLocator {
 
@@ -25,5 +29,21 @@ public class ServiceLocator {
 
     public ConfigurationAPI getConfigurationAPI() {
         return registry.resolveClass(ConfigurationAPI.class);
+    }
+
+    public SimpleLogger getSimpleLogger() {
+        return registry.resolveClass(SimpleLogger.class);
+    }
+
+    public DatabaseAPI getDatabaseAPI() {
+        return registry.resolveClass(DatabaseAPI.class);
+    }
+
+    public DataSource getDataSource() {
+        return registry.resolveClass(DataSource.class);
+    }
+
+    public Redis getRedis() {
+        return registry.resolveClass(Redis.class);
     }
 }